==> scripts/02b-run-sims-var-design.Rmd <==
---
title: "Run Precision Simulations With Design-Based Covariate Adjustment Varying P"
author: "Charlotte Mann"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(reticulate)
library(tidyr)
library(dplyr)
library(randomForest)
library(ggplot2)
library(synthpop)
library(VGAM)
library(foreach)
library(doParallel)
library(parallel)
library(devtools)
library(tibble)
#install loop from github
if(!("loop.estimator" %in% names(installed.packages()[,3]))){
  install_github("adamSales/rebarLoop/package/loop.estimator")
}
library(loop.estimator)

source("01-sim-funs.R")
source("00-helper-funs.R")

ndat <- 100
nsimt <- 1000
```

## P = 10

Run simulations in parallel. Simulate 100 new datasets (RCT and obs) and for each simulation 1,000 treatment effect vectors.

```{r p.10}
p = 10

lbound <- c(-2,rep(-2.5, p))
ubound <- c(3, rep( 2.5, p))

numCores <- detectCores()
registerDoParallel(numCores)

system.time({
  sim.list.10 <- sim_wrapper(n.obs = 10000, n.rct = 100,
                             P = p, prop.norm=1, prop.sig=.6, alpha = .5, beta.bank=.7, unex.sig = sqrt(.3),
                             lbounddp = lbound, ubounddp = ubound,
                             lambda=1, mechanism="gaussian", epsilon=c(1,3,6,10,15,30,50), delta = 10^-5,
                             n.sim.dat=ndat, n.sim.treat=nsimt, tau=.5,
                             comparison = "reloop")
})
stopImplicitCluster()
save(sim.list.10, file = "sim-out/var-design/sim-result-design-raw-10.Rdata")
```

Summarize the simulation data into a data frame rather than a list. Each row is a estimator and iteration combination, with 100 iterations of new data generations.

```{r p.10.sum}
out.p10 <- foreach(i = 1:length(sim.list.10), .combine=rbind) %do% {
  
  it.result <- sim.list.10[[i]]$result.dat %>%
    data.frame() %>%
    select(-r2) %>%
    pivot_longer(everything(), names_to = "mod", values_to = "val") %>%
    group_by(mod) %>%
    summarize(est_mean = mean(val, na.rm = T), est_var = var(val, na.rm = T)) %>%
    ungroup() %>%
    mutate(RE_sd = est_var[mod == "dif.mean"]/est_var, RE_reg = est_var[mod == "reg.est"]/est_var)
  
  it.result$reg.est.r2 <- mean(sim.list.10[[i]]$result.dat[,"r2"], na.rm = T)
  it.result$obs.r2 <- sim.list.10[[i]]$obs.r2
  it.result$rct.r2 <- sim.list.10[[i]]$rct.r2
  it.result$obs.yvar <- var(sim.list.10[[i]]$obs.y)
  it.result$rct.yvar <- var(sim.list.10[[i]]$rct.y)
  
  it.result
}

gfnorm.p10 <- foreach(i = 1:length(sim.list.10), .combine=rbind) %do% {
 
  it.result <- sim.list.10[[i]]$G.fnorm
  
  it.result
  
}

gspectral.p10 <- foreach(i = 1:length(sim.list.10), .combine=rbind) %do% {
 
  it.result <- sim.list.10[[i]]$G.spectralnorm
  
  it.result
  
}

betapred.p10 <- foreach(i = 1:length(sim.list.10), .combine=rbind) %do% {
  
  it.result <- sim.list.10[[i]]$beta.pred
  
  it.result
  
}

save(out.p10, gfnorm.p10, gspectral.p10, betapred.p10, file = "sim-out/var-design/sim-result-design-10.Rdata")
```


## P = 20

```{r p.20}
p = 20

lbound <- c(-2,rep(-2.5, p))
ubound <- c(3, rep( 2.5, p))

numCores <- detectCores()
registerDoParallel(numCores)

system.time({
  sim.list.20 <- sim_wrapper(n.obs = 10000, n.rct = 100,
                             P = p, prop.norm=1, prop.sig=.6, alpha = .5, beta.bank=.7, unex.sig = sqrt(.3),
                             lbounddp = lbound, ubounddp = ubound,
                             lambda=1, mechanism="gaussian", epsilon=c(1,3,6,10,15,30,50), delta = 10^-5,
                             n.sim.dat=ndat, n.sim.treat=nsimt, tau=.5,
                             comparison = "reloop")
})
stopImplicitCluster()
save(sim.list.20, file = "sim-out/var-design/sim-result-design-raw-20.Rdata")
```

```{r p.20.sum}
out.p20 <- foreach(i = 1:length(sim.list.20), .combine=rbind) %do% {
  
  it.result <- sim.list.20[[i]]$result.dat %>%
    data.frame() %>%
    select(-r2) %>%
    pivot_longer(everything(), names_to = "mod", values_to = "val") %>%
    group_by(mod) %>%
    summarize(est_mean = mean(val, na.rm = T), est_var = var(val, na.rm = T)) %>%
    ungroup() %>%
    mutate(RE_sd = est_var[mod == "dif.mean"]/est_var, RE_reg = est_var[mod == "reg.est"]/est_var)
  
  it.result$reg.est.r2 <- mean(sim.list.20[[i]]$result.dat[,"r2"], na.rm = T)
  it.result$obs.r2 <- sim.list.20[[i]]$obs.r2
  it.result$rct.r2 <- sim.list.20[[i]]$rct.r2
  it.result$obs.yvar <- var(sim.list.20[[i]]$obs.y)
  it.result$rct.yvar <- var(sim.list.20[[i]]$rct.y)
  
  it.result
}

gfnorm.p20 <- foreach(i = 1:length(sim.list.20), .combine=rbind) %do% {
 
  it.result <- sim.list.20[[i]]$G.fnorm
  
  it.result
  
}

gspectral.p20 <- foreach(i = 1:length(sim.list.20), .combine=rbind) %do% {
 
  it.result <- sim.list.20[[i]]$G.spectralnorm
  
  it.result
  
}

betapred.p20 <- foreach(i = 1:length(sim.list.20), .combine=rbind) %do% {
  
  it.result <- sim.list.20[[i]]$beta.pred
  
  it.result
  
}


save(out.p20, gfnorm.p20, gspectral.p20, betapred.p20, file = "sim-out/var-design/sim-result-design-20.Rdata")
```

## P = 50

```{r p.50}
p = 50

lbound <- c(-2,rep(-2.5, p))
ubound <- c(3, rep( 2.5, p))

numCores <- detectCores()
registerDoParallel(numCores)

system.time({
  sim.list.50 <- sim_wrapper(n.obs = 10000, n.rct = 100,
                             P = p, prop.norm=1, prop.sig=.6, alpha = .5, beta.bank=.7, unex.sig = sqrt(.3),
                             lbounddp = lbound, ubounddp = ubound,
                             lambda=1, mechanism="gaussian", epsilon=c(1,3,6,10,15,30,50), delta = 10^-5,
                             n.sim.dat=ndat, n.sim.treat=nsimt, tau=.5,
                             comparison = "reloop")
})
stopImplicitCluster()
save(sim.list.50, file = "sim-out/var-design/sim-result-design-raw-50.Rdata")
```

```{r p.50.sum}
out.p50 <- foreach(i = 1:length(sim.list.50), .combine=rbind) %do% {
  
  it.result <- sim.list.50[[i]]$result.dat %>%
    data.frame() %>%
    select(-r2) %>%
    pivot_longer(everything(), names_to = "mod", values_to = "val") %>%
    group_by(mod) %>%
    summarize(est_mean = mean(val, na.rm = T), est_var = var(val, na.rm = T)) %>%
    ungroup() %>%
    mutate(RE_sd = est_var[mod == "dif.mean"]/est_var, RE_reg = est_var[mod == "reg.est"]/est_var)
  
  it.result$reg.est.r2 <- mean(sim.list.50[[i]]$result.dat[,"r2"], na.rm = T)
  it.result$obs.r2 <- sim.list.50[[i]]$obs.r2
  it.result$rct.r2 <- sim.list.50[[i]]$rct.r2
  it.result$obs.yvar <- var(sim.list.50[[i]]$obs.y)
  it.result$rct.yvar <- var(sim.list.50[[i]]$rct.y)
  
  it.result
}

gfnorm.p50 <- foreach(i = 1:length(sim.list.50), .combine=rbind) %do% {
 
  it.result <- sim.list.50[[i]]$G.fnorm
  
  it.result
  
}

gspectral.p50 <- foreach(i = 1:length(sim.list.50), .combine=rbind) %do% {
 
  it.result <- sim.list.50[[i]]$G.spectralnorm
  
  it.result
  
}

betapred.p50 <- foreach(i = 1:length(sim.list.50), .combine=rbind) %do% {
  
  it.result <- sim.list.50[[i]]$beta.pred
  
  it.result
  
}


save(out.p50, gfnorm.p50, gspectral.p50, betapred.p50, file = "sim-out/var-design/sim-result-design-50.Rdata")
```
