==> scripts/02b-run-sims-var-m.Rmd <==
---
title: "Run Precision Simulations with Regression Estimator for P = 10 with Large Auxiliary Studies"
author: "Charlotte Mann"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(reticulate)
library(tidyr)
library(dplyr)
library(randomForest)
library(ggplot2)
library(synthpop)
library(VGAM)
library(foreach)
library(doParallel)
library(parallel)
library(tibble)

source("01-sim-funs.R")
source("00-helper-funs.R")

numCores <- detectCores()
```

Simulation parameters
```{r}
nobs.range <- c(1000, 10000, 50000, 100000, 250000, 500000, 1000000)
ndat <- 100
nsimt <- 1000
```


# P = 10

Provide upper and lower bounds for data columns
```{r}
p = 10

lbound <- c(-2,rep(-2.5, p))
ubound <- c(3, rep( 2.5, p))
```

```{r}
registerDoParallel(numCores)

system.time({
  sim.list.m <- lapply(nobs.range, function(x){sim_wrapper(n.obs = x, n.rct = 100, P = p, 
                                                           prop.norm=1, prop.sig=.6, alpha = .5, beta.bank=.7, unex.sig = sqrt(.3),
                                                           lbounddp = lbound, ubounddp = ubound,
                                                           lambda=1, mechanism="gaussian", epsilon=c(1,3,6,10,15,30,50), delta = 10^-5,
                                                           n.sim.dat=ndat, n.sim.treat=nsimt, tau=.5,
                                                           syn = F, comparison = "regression")})
})
stopImplicitCluster()

names(sim.list.m) <- paste("m=", nobs.range/1000,"K")
save(sim.list.m, file = "sim-out/var-auxsize/sim-result-raw-10.Rdata")
```

Summarize the simulation data into a data frame rather than a list. Each row is a estimator and iteration combination, with 100 iterations of new data generations.

```{r}
auxsizeout.p10 <- foreach(m = names(sim.list.m), .combine = rbind) %do%{
  
  tempout <- foreach(i = 1:length(sim.list.m[[m]]), .combine=rbind) %do% {
  
    it.result <- sim.list.m[[m]][[i]]$result.dat %>%
      data.frame() %>%
      select(-r2) %>%
      pivot_longer(everything(), names_to = "mod", values_to = "val") %>%
      group_by(mod) %>%
      summarize(est_mean = mean(val, na.rm = T), est_var = var(val, na.rm = T)) %>%
      ungroup() %>%
      mutate(RE_sd = est_var[mod == "dif.mean"]/est_var, RE_reg = est_var[mod == "reg.est"]/est_var)
    
    it.result$reg.est.r2 <- mean(sim.list.m[[m]][[i]]$result.dat[,"r2"], na.rm = T)
    it.result$obs.r2 <- sim.list.m[[m]][[i]]$obs.r2
    it.result$rct.r2 <- sim.list.m[[m]][[i]]$rct.r2
    it.result$obs.yvar <- var(sim.list.m[[m]][[i]]$obs.y)
    it.result$rct.yvar <- var(sim.list.m[[m]][[i]]$rct.y)
    
    it.result
  }
  
  tempout$m = m
  
  tempout
  
}


gmse.auxsize.p10 <-  foreach(m = names(sim.list.m), .combine = rbind) %do%{
  
  tempout <- foreach(i = 1:length(sim.list.m[[m]]), .combine=rbind) %do% {
 
  it.result <- sim.list.m[[m]][[i]]$G.mse
  
  it.result
  
  }
  
  c(tempout, m)
}

gspectral.auxsize.p10 <- foreach(m = names(sim.list.m), .combine = rbind) %do%{
  
  tempout = foreach(i = 1:length(sim.list.m[[m]]), .combine=rbind) %do% {
 
  it.result <- sim.list.m[[m]][[i]]$G.spectralnorm
  
  it.result
  
  }

  c(tempout, m)
}

  
save(auxsizeout.p10, gmse.auxsize.p10, gspectral.auxsize.p10, file = "sim-out/var-auxsize/sim-result-10.Rdata")
```
