==> scripts/03-figures-var.Rmd <==
---
title: "Figures - Estimating SATE with Regression Estimator"
author: "Charlotte Mann"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyr)
library(dplyr)
library(kableExtra)
library(ggplot2)
library(latex2exp)
library(RColorBrewer)
library(xtable)
library(stringr)
library(tikzDevice)
library(readr)
library(foreach)
library(tibble)
```



## Main Results

Order the estimators for tables and plots.

```{r}
mod.order <- c("dif.mean", "reg.est",
               "re.g", 
               "re.gdp1", "re.gdp3", "re.gdp6",
               "re.gdp10","re.gdp15",  "re.gdp30","re.gdp50",
               "re.syn")

mod.names <- c("No Covs", "RCT Covs", 
               "G",
               "G*, e=1", "G*, e=3", "G*, e=6",
               "G*, e=10", "G*, e=15","G*, e=30","G*, e=50",
               "Synthetic")
```


Load datasets from GreatLakes.

```{r}
load("sim-out/sim-result-10.Rdata")
load("sim-out/sim-result-20.Rdata")
load("sim-out/sim-result-50.Rdata")
```

### Relative Efficiency Table Format
#### P = 10

MSE of gram matrix for different noise additions:
```{r, echo=F}
apply(gfnorm.p10,2,mean)
```
Table of estimated variances:
```{r p10, echo = F}
plot.dat <- out.p10 %>%
  mutate(RE_sd = 1/RE_sd, RE_reg = 1/RE_reg) %>%
  group_by(mod) %>%
  summarize(across(everything(), list(mean = mean, se = ~sd(.x)/sqrt(100))))

table_1 <- plot.dat %>%
  select(mod, est_mean = est_mean_mean, est_var = est_var_mean, RE_sd = RE_sd_mean, RE_reg = RE_reg_mean, RE_sd_se, RE_reg_se) %>%
  mutate(sd_var = case_when(mod == "dif.mean" ~ est_var,
                                TRUE ~ NA_real_),
         reg_var = case_when(mod == "reg.est" ~ est_var,
                                TRUE ~ NA_real_)) %>%
  ungroup() %>%
  mutate(sd_var = mean(sd_var, na.rm = T),
         reg_var = mean(reg_var, na.rm = T)) %>%
  mutate(RE_sd2 = est_var/sd_var, RE_reg2 = est_var/reg_var) %>%
  mutate(RE_sd = case_when(RE_sd_se > .015 ~ sprintf("%1.2f (%1.2f)", RE_sd, RE_sd_se),
                           TRUE ~ sprintf("%1.2f", RE_sd)),
         RE_reg = case_when(RE_reg_se > .015 ~ sprintf("%1.2f (%1.2f)", RE_reg, RE_reg_se),
                           TRUE ~ sprintf("%1.2f", RE_reg)))

table_1$mod <- factor(table_1$mod, levels = mod.order)
table_1 <- table_1[order(table_1$mod),]

table_1 %>% 
  select(mod, est_var, RE_sd, RE_reg) %>%
  knitr::kable(format = "markdown", digits =3,
               col.names = c("Estimator", "Var","RE v SD", "RE v Reg"))
```

```{r}
xtab <- table_1 %>%
  select(est_var, RE_sd, RE_reg) %>%
  xtable(caption = "", table.placement = "ht",
               digits = c(0,3,0,0))
print(xtab, comment = F, size="footnotesize",
      file='../figures/var-sim-reg-p10.tex', include.rownames = F)
```

####  P = 20

MSE of gram matrix for different noise additions:
```{r, echo=F}
apply(gfnorm.p20,2,mean)
```
Table of estimated variances:
```{r p20,echo=F}
plot.dat <- out.p20 %>%
  mutate(RE_sd = 1/RE_sd, RE_reg = 1/RE_reg) %>%
  group_by(mod) %>%
  summarize(across(everything(), list(mean = mean, se = ~sd(.x)/sqrt(100))))

table_1 <- plot.dat %>%
  select(mod, est_mean = est_mean_mean, est_var = est_var_mean, RE_sd = RE_sd_mean, RE_reg = RE_reg_mean, RE_sd_se, RE_reg_se) %>%
  mutate(sd_var = case_when(mod == "dif.mean" ~ est_var,
                                TRUE ~ NA_real_),
         reg_var = case_when(mod == "reg.est" ~ est_var,
                                TRUE ~ NA_real_)) %>%
  ungroup() %>%
  mutate(sd_var = mean(sd_var, na.rm = T),
         reg_var = mean(reg_var, na.rm = T)) %>%
  mutate(RE_sd2 = est_var/sd_var, RE_reg2 = est_var/reg_var) %>%
  mutate(RE_sd = case_when(RE_sd_se > .015 ~ sprintf("%1.2f (%1.2f)", RE_sd, RE_sd_se),
                           TRUE ~ sprintf("%1.2f", RE_sd)),
         RE_reg = case_when(RE_reg_se > .015 ~ sprintf("%1.2f (%1.2f)", RE_reg, RE_reg_se),
                           TRUE ~ sprintf("%1.2f", RE_reg)))

table_1$mod <- factor(table_1$mod, levels = mod.order)
table_1 <- table_1[order(table_1$mod),]

table_1 %>% 
  select(mod, est_var, RE_sd, RE_reg) %>%
  knitr::kable(format = "markdown", digits =3,
               col.names = c("Estimator", "Var","RE v SD", "RE v Reg"))
```

```{r}
xtab <- table_1 %>%
  select(est_var, RE_sd, RE_reg) %>%
  xtable(caption = "", table.placement = "ht",
               digits = c(0,3,0,0))
print(xtab, comment = F, size="footnotesize",
      file='../figures/var-sim-reg-p20.tex', include.rownames = F)
```


#### P = 50

MSE of gram matrix for different noise additions:
```{r, echo=F}
apply(gfnorm.p50,2,mean)
```
Table of estimated variances:
```{r p50,echo=F}
plot.dat <- out.p50 %>%
  mutate(RE_sd = 1/RE_sd, RE_reg = 1/RE_reg) %>%
  group_by(mod) %>%
  summarize(across(everything(), list(mean = mean, se = ~sd(.x)/sqrt(100))))

table_1 <- plot.dat %>%
  select(mod, est_mean = est_mean_mean, est_var = est_var_mean, RE_sd = RE_sd_mean, RE_reg = RE_reg_mean, RE_sd_se, RE_reg_se) %>%
  mutate(sd_var = case_when(mod == "dif.mean" ~ est_var,
                                TRUE ~ NA_real_),
         reg_var = case_when(mod == "reg.est" ~ est_var,
                                TRUE ~ NA_real_)) %>%
  ungroup() %>%
  mutate(sd_var = mean(sd_var, na.rm = T),
         reg_var = mean(reg_var, na.rm = T)) %>%
  mutate(RE_sd2 = est_var/sd_var, RE_reg2 = est_var/reg_var) %>%
  mutate(RE_sd = case_when(RE_sd_se > .015 ~ sprintf("%1.2f (%1.2f)", RE_sd, RE_sd_se),
                           TRUE ~ sprintf("%1.2f", RE_sd)),
         RE_reg = case_when(RE_reg_se > .015 ~ sprintf("%1.2f (%1.2f)", RE_reg, RE_reg_se),
                           TRUE ~ sprintf("%1.2f", RE_reg)))

table_1$mod <- factor(table_1$mod, levels = mod.order)
table_1 <- table_1[order(table_1$mod),]

table_1 %>% 
  select(mod, est_var, RE_sd, RE_reg) %>%
  knitr::kable(format = "markdown", digits =3,
               col.names = c("Estimator", "Var","RE v SD", "RE v Reg"))
```

```{r}
xtab <- table_1 %>%
  select(est_var, RE_sd, RE_reg) %>%
  xtable(caption = "", table.placement = "ht",
               digits = c(0,3,0,0))
print(xtab, comment = F, size="footnotesize",
      file='../figures/var-sim-reg-p50.tex', include.rownames = F)
```

### Relative Efficiency Plot Format

```{R}
all.out <- out.p10 %>%
  mutate(p = "P = 10") %>%
  bind_rows(out.p20) %>%
  mutate(p = case_when(is.na(p) ~ "P = 20",
                       TRUE ~ p)) %>%
  bind_rows(out.p50) %>%
  mutate(p = case_when(is.na(p) ~ "P = 50",
                        TRUE ~ p)) %>%
  select(mod, p, est_mean, est_var, RE_sd, RE_reg) %>%
  filter(mod %in% mod.order) %>%
  group_by(mod, p) %>%
  summarize(across(everything(), list(mean = mean, se = ~sd(.x)/sqrt(100))))

plot.dat <- all.out %>%
  select(mod,p, est_mean = est_mean_mean, est_var = est_var_mean, RE_sd = RE_sd_mean, RE_reg = RE_reg_mean, RE_sd_se, RE_reg_se) %>%
  ungroup() %>%
  mutate(sd_var = case_when(mod == "dif.mean" ~ est_var,
                                TRUE ~ NA_real_),
         reg_var = case_when(mod == "reg.est" ~ est_var,
                                TRUE ~ NA_real_)) %>%
  group_by(p) %>%
  mutate(sd_var = mean(sd_var, na.rm = T),
         reg_var = mean(reg_var, na.rm = T)) %>%
  mutate(RE_sd2 = sd_var/est_var, RE_reg2 = reg_var/est_var) %>%
  mutate(ub = RE_reg + 2*RE_reg_se, lb = RE_reg - 2*RE_reg_se) %>%
  mutate(type = case_when(str_detect(mod,"gdp") ~ "GDP",
                          mod %in% c("dif.mean","reg.est") ~ "RCT",
                          str_detect(mod,"syn") ~ "synth",
                          mod == "re.gtilde" ~ "gtilde",
                          TRUE ~ "g"
                          )) %>%
  arrange(p, RE_reg) %>%
  group_by(p) %>%
  mutate(id = row_number())

plot.dat$mod <- factor(plot.dat$mod, levels = mod.order)
```



```{r}
cols <- c("#fd8d3c","#f768a1",
          "#6e016b",
          "#a1d99b","#74c476","#41ab5d","#238b45","#005a32","#00441b", "black",
          "#4292c6")
 
shapes <- c(19, 19,
            17,
            17,17,17,17,17,17,17,
            15)
```


```{r}
g <- ggplot(data = plot.dat) +
  geom_hline(yintercept = 1) +
  geom_errorbar(aes(x = id, ymin = lb, ymax = ub), width = .5, linewidth = .2)+
  geom_point(aes(x = id, y = RE_reg, color = mod, shape = mod), size =1.5)+ 
  facet_wrap(~p)+
  xlab("") + 
  scale_color_manual(name="", values = cols, labels = mod.names)+
  scale_shape_manual(name="", values = shapes, labels = mod.names) +
  scale_y_log10(name = "Sample Size Multiplier\n(compared with adjusting with RCT covariates only)", 
                breaks =c(.3,.5,.75,1,1.25, 1.5, 1.75, 2, 2.25), limits = c(.3, 2.3)) +
  #scale_y_continuous(name="Relative Efficiency", breaks =c(.5,.75,1,1.25, 1.5, 1.75, 2, 2.25), limits=c(.3,2.27))+
  # theme(text = element_text(family = "sans",
  #                           size = 10),
  #       panel.border = element_blank(),
  #       panel.grid.major.x = element_blank(),
  #       panel.grid.minor = element_blank(),
  #       #panel.grid.major.y = 
  #       plot.title = element_text(hjust = 0.5),
  #       axis.text.x = element_blank(),
  #       axis.text.y = element_text(size = 8),
  #       axis.title.y = element_text(size = 8),
  #       axis.title.x = element_text(size = 8),
  #       strip.text = element_text(size = 10),
  #       legend.position = "right",
  #       legend.justification="top",
  #       axis.ticks.x = element_blank(),
  #       legend.box.spacing = unit(0, 'cm'),
  #       legend.text = element_text(size = 8),
  #       legend.margin=ggplot2::margin(-10,0,0,7),
  #       plot.margin = unit(c(0, 0, 0, 0),"inches"))
 theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major.y = element_line(color = "grey80"),
        #axis.line.x = element_line(color = "grey10"),
        #axis.line.y = element_line(color = "black"),
        axis.ticks.x = element_blank(),
        axis.text.x = element_blank(),
        plot.title = element_text(hjust = 0.5),
        axis.text.y = element_text(size = 8),
        axis.title.y = element_text(size = 10),
        axis.title.x = element_text(size = 8),
        strip.text = element_text(size = 10),
        panel.background = element_rect(fill='white'),
        legend.position = "right",
        legend.justification="top",
        legend.box.spacing = unit(0, 'cm'),
        legend.text = element_text(size = 8),
        legend.margin=ggplot2::margin(0,0,0,7),
        plot.margin = unit(c(0, 0, 0, 0),"inches"))

g
```


```{r}
tikz(file = "../figures/re-regression-paper.tex", width = 5.5, height = 4, symbolicColors = T, colorFileName = "../figures/figcolors.tex")
g
dev.off()
```

### Privacy Output Metrics

```{r}
betalist <- list("10" = betapred.p10,"20" = betapred.p20,"50" = betapred.p50)
spectrallist <- list("10" = gspectral.p10, "20" = gspectral.p20, "50" = gspectral.p50)
gfnormlist <- list("10" = gfnorm.p10, "20" = gfnorm.p20, "50" = gfnorm.p50)
```


MSE of outcome model coefficients:

```{r}
beta.mse.all <- foreach(p = names(betalist), .combine = rbind) %do% {
  beta.mse <- betalist[[p]] %>%
    rownames_to_column(var = "coef") %>%
    mutate(it = case_when(str_detect(coef, "Intercept") & str_detect(coef, "\\d") ~ str_extract(coef, "\\d+"),
                          str_detect(coef, "Intercept") ~ "0",
                          TRUE ~ NA)) %>%
    fill(it) %>%
    mutate(across(!it & !coef, .fns = ~(.x - true)^2,.names = "{.col}_err")) %>%
    group_by(it) %>%
    summarize(across(ends_with("err"), mean)) %>%
    pivot_longer(!it, names_to = "privmethod", values_to = "mse") %>%
    group_by(privmethod) %>%
    summarize(e_rmse = mean(sqrt(mse)), se_rmse = sd(sqrt(mse))/sqrt(100)) %>%
    mutate(privmethod = str_replace(privmethod, "\\_err", "")) %>%
    filter(privmethod != "true") %>%
    mutate(p = paste0("P = ",p))
}
```

Frobenius norm of gram matrix - non-private gram matrix

```{r}
gfnorm.all <- foreach(p = names(gfnormlist), .combine = rbind) %do% {
  gram.mse <- gfnormlist[[p]] %>%
    data.frame() %>%
    rownames_to_column(var = "it") %>%
    pivot_longer(!it, names_to = "privmethod", values_to = "mse") %>%
    group_by(privmethod) %>%
    summarize(e_gfnorm = mean(mse), se_gfnorm = sd(mse)/sqrt(100)) %>%
    mutate(privmethod = str_remove_all(privmethod, "\\.")) %>%
    mutate(privmethod = str_replace(privmethod, "gsyn", "syn")) %>%
    mutate(p = paste0("P = ",p))
}
```


Spectral norm of gram matrix - nonprivate gram matrix

```{r}
gspectral.all <- foreach(p = names(spectrallist), .combine = rbind) %do% {
  gram.spectral <- spectrallist[[p]] %>%
    data.frame() %>%
    rownames_to_column(var = "it") %>%
    pivot_longer(!it, names_to = "privmethod", values_to = "norm") %>%
    group_by(privmethod) %>%
    summarize(e_norm = mean(norm), norm_se = sd(norm)/sqrt(100)) %>%
    mutate(privmethod = str_remove_all(privmethod, "\\.")) %>%
    mutate(privmethod = str_replace(privmethod, "gsyn", "syn")) %>%
    mutate(p = paste0("P = ",p))
}
```


Combining all utility metrics

```{r}
met.order <- c("gram" , "syn", "epsilon1", "epsilon3", "epsilon6", "epsilon15", "epsilon30", "epsilon50")
table.lables <- c("$\\g$ (No privacy)", "$\\syn$", "$\\gstar$ (DP)", "$\\gstar$ (DP)", "$\\gstar$ (DP)","$\\gstar$ (DP)","$\\gstar$ (DP)","$\\gstar$ (DP)")

xwalk <- data.frame(privmethod = met.order, labs = table.lables)
```

```{r}
utility.metrics <- beta.mse.all %>%
  left_join(gfnorm.all, by = c("privmethod", "p")) %>%
  left_join(gspectral.all, by = c("privmethod", "p")) %>%
  filter(privmethod %in% met.order) %>%
  mutate(privmethod = factor(privmethod, levels = met.order)) %>%
  select( p, privmethod, everything()) %>%
  arrange(p,privmethod) %>%
  mutate(across(contains("norm"), ~ case_when(is.na(.x) ~ 0, TRUE ~ .x))) %>%
  mutate(beta_rmse = case_when(se_rmse > .0044 ~ sprintf("%1.2f (%1.2f)", e_rmse, se_rmse),
                                  TRUE ~ sprintf("%1.2f", e_rmse)),
         gram_frob_norm =  case_when(se_gfnorm > .0044 ~ sprintf("%1.2f (%1.2f)", e_gfnorm, se_gfnorm),
                                     TRUE ~ sprintf("%1.2f", e_gfnorm)),
         gram_spec_norm =  case_when(norm_se > .0044 ~ sprintf("%1.2f (%1.2f)", e_norm, norm_se),
                                     TRUE ~ sprintf("%1.2f", e_norm))) %>%
  mutate(ep = case_when(str_detect(privmethod, "epsilon") ~ str_extract(privmethod, "\\d+"),
                                   TRUE ~ "-")) %>%
  left_join(xwalk, by = "privmethod")

utility.metrics
```


```{r}
utility.metrics %>%
  select(labs, ep, e_rmse, e_gfnorm, e_norm) %>%
  kable(booktabs = T, col.names = c("Release","$\\epsilon$" ,"RMSE", "Frobenius Norm", "Spectral Norm"),
                    escape = FALSE, align = "lcccc", full_width = F, digits =2) %>% 
  kable_styling(latex_options = c("hold_position","scale_down")) %>%
  row_spec(row = 0, align = "c", bold=T) %>%
  pack_rows("P = 10", 1,8) %>%
  pack_rows("P = 20", 9, 16) %>%
  pack_rows("P = 50", 17, 24) %>%
  add_header_above(c(" " = 2, "$\\hat{\\beta}$" = 1,"Gram Matrix" = 2),  bold = F)
```

```{r}
df <- utility.metrics %>%
  select(labs, ep, e_rmse, e_gfnorm, e_norm) %>%
  kable(booktabs = T, col.names = c("Release","$\\epsilon$" ,"RMSE", "Frobenius Norm", "Spectral Norm"),
                    escape = FALSE, align = "lcccc", full_width = F, digits =2, format = "latex") %>% 
  kable_styling(latex_options = c("hold_position","scale_down")) %>%
  row_spec(row = 0, align = "c", bold=T) %>%
  pack_rows("P = 10", 1,8) %>%
  pack_rows("P = 20", 9, 16) %>%
  pack_rows("P = 50", 17, 24) %>%
  add_header_above(c(" " = 2, "$\\hat{\\beta}$" = 1,"Gram Matrix" = 2),  bold = F)

write_lines(df, '../figures/utility_metrics_main.tex')
out <- read_lines('../figures/utility_metrics_main.tex')
out <- str_replace(out, "beta", "$\\\\hat{\\\\bm \\\\beta}$")
out <- str_replace(out, "Gram Matrix", "\\\\textbf{Gram Matrix}")
write_lines(out[-c(1, length(out))], '../figures/utility_metrics_main.tex')
```

## Increasing Auxiliary Size



```{r}
mod.order <- c("dif.mean", "reg.est",
               "re.g", 
               "re.gdp1", "re.gdp3", "re.gdp6",
               "re.gdp10","re.gdp15",  "re.gdp30","re.gdp50")

mod.names <- c("No Covs", "RCT Covs", 
               "G",
               "G*, e=1", "G*, e=3", "G*, e=6",
               "G*, e=10", "G*, e=15","G*, e=30","G*, e=50")


cols <- c("#fd8d3c","#f768a1",
          "#6e016b",
          "#a1d99b","#74c476","#41ab5d","#238b45","#005a32","#00441b", "black")
 
shapes <- c(19, 19,
            17,
            17,17,17,17,17,17,17)
```



```{r}
load("sim-out/var-auxsize/sim-result-10.Rdata")
load("sim-out/var-auxsize/sim-result-20.Rdata")
```

### P = 10

```{r}
aux.out <- auxsizeout.p10 %>%
  filter(mod %in% mod.order) %>%
  group_by(mod, m) %>%
  summarize(across(everything(), list(mean = mean, se = ~sd(.x)/sqrt(100))))

plot.dat.aux <- aux.out  %>%
  select(mod, m, est_var = est_var_mean, est_var_se) %>%
  ungroup() %>%
  mutate(ub = est_var + 2*est_var_se, lb = est_var - 2*est_var_se) %>%
  mutate(type = case_when(str_detect(mod,"gdp") ~ "GDP",
                          mod %in% c("dif.mean","reg.est") ~ "RCT",
                          mod == "re.gtilde" ~ "gtilde",
                          TRUE ~ "g")) %>%
  mutate(auxsize = as.numeric(str_extract_all(m, "\\d+"))*1000)

plot.dat.aux$mod <- factor(plot.dat.aux$mod, levels = mod.order, labels = mod.names)
```


```{r}
g <- ggplot(data = plot.dat.aux) +
  geom_errorbar(aes(x = auxsize, ymin = lb, ymax = ub), width = .05, linewidth = .2)+
  geom_point(aes(x = auxsize, y = est_var, color = mod, shape = mod), size =1.5)+ 
  geom_line(aes(x = auxsize, y = est_var, color = mod), linewidth = .5) +
  scale_color_manual(name="Covariates", values = cols)+
  scale_shape_manual(name="Covariates", values = shapes) +
  scale_x_log10(name = "Auxiliary Size (m)", breaks = c(1000, 10000, 50000, 100000, 250000, 500000, 1000000), 
                labels = scales::label_comma()) +
  ylab("Variance") +
  theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major.y = element_line(color = "grey80", linetype = 3),
        axis.line.x = element_line(color = "grey10"),
        #axis.line.y = element_line(color = "black"),
        plot.title = element_text(hjust = 0.5),
        axis.text.y = element_text(size = 8),
        axis.title.y = element_text(size = 10),
        axis.title.x = element_text(size = 10),
        strip.text = element_text(size = 10),
        panel.background = element_rect(fill='white'),
        legend.position = "right",
        legend.justification="top",
        legend.box.spacing = unit(0, 'cm'),
        legend.text = element_text(size = 8),
        legend.margin=ggplot2::margin(5,0,0,7),
        plot.margin = unit(c(0, 0, 0, 0),"inches"))

g
```

```{r}
tikz(file = "../figures/auxsize-reg.tex", width = 5.7, height = 4,
     symbolicColors = T, colorFileName = "../figures/figcolorssize.tex")
g
dev.off()
```


### P = 20

```{r}
aux.out <- auxsizeout.p20 %>%
  group_by(mod, m) %>%
  summarize(across(everything(), list(mean = mean, se = ~sd(.x)/sqrt(100))))

plot.dat.aux <- aux.out  %>%
  select(mod, m, est_var = est_var_mean, est_var_se) %>%
  ungroup() %>%
  mutate(ub = est_var + 2*est_var_se, lb = est_var - 2*est_var_se) %>%
  mutate(type = case_when(str_detect(mod,"gdp") ~ "GDP",
                          mod %in% c("dif.mean","reg.est") ~ "RCT",
                          mod == "re.gtilde" ~ "gtilde",
                          TRUE ~ "g")) %>%
  mutate(auxsize = as.numeric(str_extract_all(m, "\\d+"))*1000)

plot.dat.aux$mod <- factor(plot.dat.aux$mod, levels = mod.order, labels = mod.names)
```


```{r}
g <- ggplot(data = plot.dat.aux) +
  geom_errorbar(aes(x = auxsize, ymin = lb, ymax = ub), width = .05, linewidth = .2)+
  geom_point(aes(x = auxsize, y = est_var, color = mod, shape = mod), size =1.5)+ 
  geom_line(aes(x = auxsize, y = est_var, color = mod), linewidth = .5) +
  scale_color_manual(name="", values = cols)+
  scale_shape_manual(name="", values = shapes) +
  scale_x_log10(name = "Auxiliary Size (m)", breaks = c(1000, 10000, 50000, 100000, 250000, 500000, 1000000), 
                labels = scales::label_comma()) +
  ylab("Variance") +
  ggtitle("P = 20")
  theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major.y = element_line(color = "grey80", linetype = 3),
        axis.line.x = element_line(color = "grey10"),
        #axis.line.y = element_line(color = "black"),
        plot.title = element_text(hjust = 0.5),
        axis.text.y = element_text(size = 8),
        axis.title.y = element_text(size = 8),
        axis.title.x = element_text(size = 8),
        strip.text = element_text(size = 10),
        panel.background = element_rect(fill='white'),
        legend.position = "right",
        legend.justification="top",
        legend.box.spacing = unit(0, 'cm'),
        legend.text = element_text(size = 8),
        legend.margin=ggplot2::margin(0,0,0,7),
        plot.margin = unit(c(0, 0, 0, 0),"inches"))

g
```


## Design-Based Estimator Appendix

```{r}
load("sim-out/var-design/sim-result-design-10.Rdata")
load("sim-out/var-design/sim-result-design-20.Rdata")
load("sim-out/var-design/sim-result-design-50.Rdata")
```

```{r}
mod.order.design <- c("dif.mean", "reg.est",
               "reloop.g", 
               "reloop.gdp1", "reloop.gdp3", "reloop.gdp6",
               "reloop.gdp10","reloop.gdp15",  "reloop.gdp30","reloop.gdp50",
               "reloop.syn")
```

```{R}
all.out.design  <- out.p10 %>%
  mutate(p = "P = 10") %>%
  bind_rows(out.p20) %>%
  mutate(p = case_when(is.na(p) ~ "P = 20",
                       TRUE ~ p)) %>%
  bind_rows(out.p50) %>%
  mutate(p = case_when(is.na(p) ~ "P = 50",
                        TRUE ~ p)) %>%
  select(mod, p, est_mean, est_var, RE_sd, RE_reg) %>%
  filter(mod %in% mod.order.design) %>%
  group_by(mod, p) %>%
  summarize(across(everything(), list(mean = mean, se = ~sd(.x)/sqrt(100))))

plot.dat.design <- all.out.design  %>%
  select(mod,p, est_mean = est_mean_mean, est_var = est_var_mean, RE_sd = RE_sd_mean, RE_reg = RE_reg_mean, RE_sd_se, RE_reg_se) %>%
  mutate(ub = RE_reg + 2*RE_reg_se, lb = RE_reg - 2*RE_reg_se) %>%
  arrange(p, RE_reg) %>%
  group_by(p) %>%
  mutate(id = row_number())

plot.dat.design$mod <- factor(plot.dat.design$mod, levels = mod.order.design)
```



```{r}
cols <- c("#fd8d3c","#f768a1",
          "#6e016b",
          "#a1d99b","#74c476","#41ab5d","#238b45","#005a32","#00441b", "black",
          "#4292c6")
 
shapes <- c(19, 19,
            17,
            17,17,17,17,17,17,17,
            15)
```


```{r}
g <- ggplot(data = plot.dat.design) +
  geom_hline(yintercept = 1) +
  geom_errorbar(aes(x = id, ymin = lb, ymax = ub), width = .5, linewidth = .2)+
  geom_point(aes(x = id, y = RE_reg, color = mod, shape = mod), size =1.5)+ 
  facet_wrap(~p)+
  xlab("") + 
  scale_color_manual(name="", values = cols, labels = mod.names)+
  scale_shape_manual(name="", values = shapes, labels = mod.names) +
  scale_y_log10(name = "Sample Size Multiplier\n(compared with adjusting with RCT covariates only)", 
                breaks =c(.3,.5,.75,1,1.25, 1.5, 1.75, 2, 2.25), limits = c(.3, 2.3)) +
 theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor = element_blank(),
        panel.grid.major.y = element_line(color = "grey80"),
        #axis.line.x = element_line(color = "grey10"),
        #axis.line.y = element_line(color = "black"),
        axis.ticks.x = element_blank(),
        axis.text.x = element_blank(),
        plot.title = element_text(hjust = 0.5),
        axis.text.y = element_text(size = 8),
        axis.title.y = element_text(size = 10),
        axis.title.x = element_text(size = 8),
        strip.text = element_text(size = 10),
        panel.background = element_rect(fill='white'),
        legend.position = "right",
        legend.justification="top",
        legend.box.spacing = unit(0, 'cm'),
        legend.text = element_text(size = 8),
        legend.margin=ggplot2::margin(0,0,0,7),
        plot.margin = unit(c(0, 0, 0, 0),"inches"))

g
```


```{r}
tikz(file = "../figures/re-design-appx.tex", width = 5.5, height = 4, symbolicColors = T, colorFileName = "../figures/figcolorsdesign.tex")
g
dev.off()
```
