==> scripts/12-run-sims-gen.Rmd <==
---
title: "Run Generalizing Simulations Changing P"
author: "Charlotte Mann"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(reticulate)
library(tidyr)
library(dplyr)
library(randomForest)
library(ggplot2)
library(synthpop)
library(VGAM)
library(foreach)
library(doParallel)
library(parallel)
library(stringr)
library(nleqslv)
library(tibble)


source("00-helper-funs.R")
source("00-cw-funs.R")
source("11-sim-funs-gen.R")
```


Set some parameters for all simulations
```{r}
inference = T
dp.syn = F
n.sim = 1000
```

## P = 10

```{r,p.10}
p = 10
lbound <- c(0,rep(-1.5, p+1))
ubound <- c(5, rep( 3.5, p+1))

numCores <- detectCores()
registerDoParallel(numCores)

system.time({
  
  sim.gen.10 <- sim_wrapper_gen(n.obs=10000, n.rct=100, P=10, prop.norm=1, prop.sig=.6,
                        alpha = .5, beta.bank=.7, unex.sig = sqrt(.3),                   
                        select.alpha =-2, select.gamma=1,                    
                        select.beta.bank = -1, select.prop.sig=.5,
                        lambda=1, syn.epsilon = 3,                                        
                        mechanism = "gaussian", epsilon = c(1,3,6,15,30), delta = 10^-5, 
                        lbounddp = lbound, ubounddp = ubound,
                        n.sim.dat=n.sim, n.sim.treat=1, tau=.5,                    
                        syn.dp = dp.syn, cw.inference = inference,
                        nboot.inf = 100, level= .05)
  
})

stopImplicitCluster()

save(sim.gen.10, file = "sim-out/gen-sim-result-raw-10.Rdata")
```


```{r,p.10.save}
out.p10 <- foreach(i = 1:(length(sim.gen.10)-2), .combine=rbind) %do% {
  
  if(!is.null(sim.gen.10[[i]]$result.dat)){
    it.result <- sim.gen.10[[i]]$result.dat %>%
      mutate(it = i) %>%
      rownames_to_column(var = "type")
    
    it.result
  }
}



est.p10 <- out.p10 %>%
  filter(type == "est")

coverage.p10 <- out.p10 %>%
  filter(type == "coverage")

if(!inference){
  coverage.p10 <- coverage.p10 %>%
    select(dif.mean, reg.est, it)
}



auxmeans.p10 <- foreach(i = 1:(length(sim.gen.10)-2), .combine=rbind) %do% {
 
  it.result <- sim.gen.10[[i]]$aux.means
  
  it.result
  
}


beta.10 <- sim.gen.10$beta
select.beta.10 <- sim.gen.10$select.beta


save(est.p10, coverage.p10, auxmeans.p10, beta.10, select.beta.10,
     file = "sim-out/gen-sim-result-10.Rdata")
```

## P = 20

```{r,p.20}
p = 20
lbound <- c(1,rep(-1.5, p+1))
ubound <- c(6, rep(3.5, p+1))


registerDoParallel(numCores)

system.time({
  
  sim.gen.20 <- sim_wrapper_gen(n.obs=10000, n.rct=100, P=20, prop.norm=1, prop.sig=.6,
                        alpha = .5, beta.bank=.7, unex.sig = sqrt(.3),                   
                        select.alpha =-2, select.gamma=1,                    
                        select.beta.bank = -1, select.prop.sig=.5,
                        lambda=1, syn.epsilon = 3,                                        
                        mechanism = "gaussian", epsilon = c(3,6,15,30), delta = 10^-5, 
                        lbounddp = lbound, ubounddp = ubound,
                        n.sim.dat=n.sim, n.sim.treat=1, tau=.5,                    
                        syn.dp = dp.syn, cw.inference = inference,
                        nboot.inf = 100, level= .05)
  
})

stopImplicitCluster()

save(sim.gen.20, file = "sim-out/gen-sim-result-raw-20.Rdata")
```

```{r,p.20.save}
out.p20 <- foreach(i = 1:(length(sim.gen.20)-2), .combine=rbind) %do% {
  
  if(!is.null(sim.gen.20[[i]]$result.dat)){
    it.result <- sim.gen.20[[i]]$result.dat %>%
      mutate(it = i) %>%
      rownames_to_column(var = "type")
    
    it.result
  }
}



est.p20 <- out.p20 %>%
  filter(type == "est")

coverage.p20 <- out.p20 %>%
  filter(type == "coverage")

if(!inference){
  coverage.p20 <- coverage.p20 %>%
    select(dif.mean, reg.est, it)
}



auxmeans.p20 <- foreach(i = 1:(length(sim.gen.20)-2), .combine=rbind) %do% {
 
  it.result <- sim.gen.20[[i]]$aux.means
  
  it.result
  
}



beta.20 <- sim.gen.20$beta
select.beta.20 <- sim.gen.20$select.beta

save(est.p20, coverage.p20, auxmeans.p20, beta.20, select.beta.20,
     file = "sim-out/gen-sim-result-20.Rdata")
```
