==> scripts/13-figures-gen.Rmd <==
---
title: "Generalizabilty Results"
author: "Charlotte Mann"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
library(RColorBrewer)
library(dplyr)
library(tidyr)
library(stringr)
library(xtable)
library(tikzDevice)
library(forcats)
```

```{r}
mods <- c("dif.mean", "reg.est","acwt.aux",
          "acwt.epsilon1","acwt.epsilon3","acwt.epsilon6",
          "acwt.epsilon15","acwt.epsilon30", "acwt.syn")

mod.names <- c("Simple Difference", 
               "Regression", 
               "ACW-t - G",
               "ACW-t - G*, e=1",
               "ACW-t - G*, 3",
               "ACW-t - G*, 6",
               "ACW-t - G*, 15",
               "ACW-t - G*, 30",
               "ACW-t - Synthetic")
```
## Separate Plots

### P = 10

```{r}
load("sim-out/gen-sim-result-10.Rdata")
```

```{r}
plot.dat <- est.p10 %>%
  select(-sate, -type) %>%
  pivot_longer(!it, names_to = "model", values_to = "est") %>%  
  filter(model %in% mods)

plot.dat$mod <- factor(plot.dat$model, levels = mods, labels = mod.names)
```

```{r}
ggplot(data = plot.dat, aes(y = est, x = mod)) +
  geom_boxplot() +
  geom_abline(intercept = .5, slope = 0) +
  geom_abline(intercept = mean(est.p10$sate), slope = 0, linetype = "dashed", color= "red")+
  xlab("Estimator") + ylab("Estimate") +
  theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        axis.text.x = element_text(angle=45, size = 10, hjust =1),
        axis.text.y = element_text(size = 10),
        axis.title.y = element_text(size = 10),
        axis.title.x = element_text(size = 10),
        legend.text = element_text(size = 8))
```


```{r}
est.p10 %>%
  select(-type, -sate, -it) %>%
  summarize(across(everything(), ~mean(is.na(.x)))) %>%
  t()
```

```{r}
groups <- factor(sample(1:20, nrow(est.p10), replace = T))

sim.se <- est.p10 %>%
  mutate(group = groups) %>%
  select(-sate, -type) %>%
  pivot_longer(!(it|group), names_to = "model", values_to = "val") %>%
  group_by(model, group)%>%
  summarize(est_mean = mean(val, na.rm = T), est_var = var(val, na.rm = T)) %>%
  mutate(bias2 = (est_mean-.5)^2, mse = est_var + bias2)  %>%
  group_by(model) %>%
  summarize(mse_sim_se = sd(mse)/sqrt(20),
            group_mse = mean(mse),
            est_var_sim_se = sd(est_var)/sqrt(20),
            bias2_sim_se = sd(bias2)/sqrt(20))


plot.dat <- est.p10 %>%
  select(-sate, -type) %>%
  pivot_longer(!it, names_to = "model", values_to = "val") %>%
  group_by(model)%>%
  summarize(est_mean = mean(val, na.rm = T), est_var = var(val, na.rm = T)) %>%
  mutate(bias2 = (est_mean-.5)^2, mse = bias2 + est_var) %>%
  left_join(sim.se, by = "model") %>%
  mutate(ub_mse = mse + 2*mse_sim_se,
         lb_mse = mse - 2*mse_sim_se) %>%
  filter(model %in% mods) %>%
  mutate(model = factor(model,  levels = mods, labels = mod.names))
```


```{r}
plot.dat.long <- plot.dat %>%
  select(model,est_var, bias2) %>%
  pivot_longer(est_var:bias2, names_to = "component",
               values_to = "value")
```


```{r}
cols <- c("#FDBF6F","#1F78B4")

g <- ggplot() +
  geom_bar(position="stack", stat="identity", data = plot.dat.long, aes(fill = component, y = value, x = model)) +
  geom_errorbar(data = plot.dat, aes(x = model, ymin = lb_mse, ymax = ub_mse), linewidth = .4, width = .5) +
  ylab("MSE") + xlab("Estimator") +
  ggtitle("P = 10")+
  scale_fill_manual(name="", values= cols,labels =c("Bias","Variance"))+
  theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        plot.title = element_text(hjust = 0.5),
        axis.text.x = element_text(angle=45, size = 10, hjust =1),
        axis.text.y = element_text(size = 10),
        axis.title.y = element_text(size = 10),
        axis.title.x = element_text(size = 10),
        legend.text = element_text(size = 10))
  
g
```

```{r}
png("../figures/gen-mse-p10.png", width = 6, height = 4.5, units = 'in', res = 300)
g
dev.off()
```


```{R, sate}
mean(est.p10$sate)
```

### P = 20

```{r}
load("sim-out/gen-sim-result-20.Rdata")
```

```{r}
plot.dat1 <- est.p20 %>%
  select(-sate, -type) %>%
  pivot_longer(!it, names_to = "model", values_to = "est") %>%  
  filter(model %in% mods)

plot.dat1$mod <- factor(plot.dat1$model, levels = mods, labels = mod.names)
```

```{r}
ggplot(data = plot.dat1, aes(y = est, x = mod)) +
  geom_boxplot() +
  geom_abline(intercept = .5, slope = 0) +
  geom_abline(intercept = mean(est.p20$sate), slope = 0, linetype = "dashed", color= "red")+
  xlab("Estimator") + ylab("Estimate") +
  theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        axis.text.x = element_text(angle=45, size = 10, hjust =1),
        axis.text.y = element_text(size = 10),
        axis.title.y = element_text(size = 10),
        axis.title.x = element_text(size = 10),
        legend.text = element_text(size = 8))
```

```{r}
est.p20 %>%
  select(-type, -sate, -it) %>%
  summarize(across(everything(), ~mean(is.na(.x)))) %>%
  t()
```

```{r}
groups <- factor(sample(1:20, nrow(est.p20), replace = T))

sim.se20 <- est.p20 %>%
  mutate(group = groups) %>%
  select(-sate, -type) %>%
  pivot_longer(!(it|group), names_to = "model", values_to = "val") %>%
  group_by(model, group)%>%
  summarize(est_mean = mean(val, na.rm = T), est_var = var(val, na.rm = T)) %>%
  mutate(bias2 = (est_mean-.5)^2, mse = est_var + bias2)  %>%
  group_by(model) %>%
  summarize(mse_sim_se = sd(mse)/sqrt(20),
            group_mse = mean(mse),
            est_var_sim_se = sd(est_var)/sqrt(20),
            bias2_sim_se = sd(bias2)/sqrt(20))


plot.dat20 <- est.p20 %>%
  select(-sate, -type) %>%
  pivot_longer(!it, names_to = "model", values_to = "val") %>%
  group_by(model)%>%
  summarize(est_mean = mean(val, na.rm = T), est_var = var(val, na.rm = T)) %>%
  mutate(bias2 = (est_mean-.5)^2, mse = bias2 + est_var) %>%
  left_join(sim.se20, by = "model") %>%
  mutate(ub_mse = mse + 2*mse_sim_se,
         lb_mse = mse - 2*mse_sim_se) %>%
  filter(model %in% mods) %>%
  mutate(model = factor(model,  levels = mods, labels = mod.names))
```


```{r}
plot.dat.long20 <- plot.dat20 %>%
  select(model,est_var, bias2) %>%
  pivot_longer(est_var:bias2, names_to = "component",
               values_to = "value")
```


```{r}
cols <- c("#FDBF6F","#1F78B4")

g <- ggplot() +
  geom_bar(position="stack", stat="identity", data = plot.dat.long20, aes(fill = component, y = value, x = model)) +
  geom_errorbar(data = plot.dat20, aes(x = model, ymin = lb_mse, ymax = ub_mse), linewidth = .4, width = .5) +
  ylab("MSE") + xlab("Estimator") +
  ggtitle("P = 20")+
  scale_fill_manual(name="", values= cols,labels =c("Bias","Variance"))+
  theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        plot.title = element_text(hjust = 0.5),
        axis.text.x = element_text(angle=45, size = 10, hjust =1),
        axis.text.y = element_text(size = 10),
        axis.title.y = element_text(size = 10),
        axis.title.x = element_text(size = 10),
        legend.text = element_text(size = 10))
  
g
```

```{r}
png("../figures/gen-mse-p20.png", width = 6, height = 4.5, units = 'in', res = 300)
g
dev.off()
```

## Combined plot

```{r}
plot.dat.long20$ncovs = "p = 20"
plot.dat.long.combined <- plot.dat.long %>%
  mutate(ncovs = "p = 10") %>% 
  rbind(plot.dat.long20) %>%
  mutate(component = factor(component, levels = c("est_var","bias2")))

plot.dat20$ncovs = "p = 20"
plot.dat.combined <- plot.dat %>%
  mutate(ncovs = "p = 10") %>% 
  rbind(plot.dat20)
```

```{r}
cols <- c("#1F78B4","#FDBF6F")

g <- ggplot() +
  geom_bar(position="stack", stat="identity",data = plot.dat.long.combined, aes(fill = component, y = value, x = fct_rev(model))) +
   geom_errorbar(data = plot.dat.combined, aes(x = fct_rev(model), ymin = lb_mse, ymax = ub_mse), linewidth = .4, width = .5) +
  facet_wrap(~ncovs)+
  ylab("") + 
  xlab("") +
  coord_flip() +
  scale_fill_manual(name="MSE = ", values= cols,labels =c("Variance + ", "Squared Bias"))+
  #ggtitle("P = 20")+
  theme(text = element_text(family = "sans",
                            size = 10),
        panel.border = element_blank(),
        panel.background = element_rect(fill='white'),
        panel.grid.major.x = element_line(color = "grey80"),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        plot.title = element_text(hjust = 0.5),
        axis.title = element_blank(),
        axis.text.y = element_text(size = 10,color = "black"),
        axis.text.x = element_text(size = 8),
        axis.line.x = element_line(color = "black"),
        #axis.line.y = element_line(color = "black"),
        strip.text = element_text(size = 10),
        legend.position = "bottom",
        legend.text = element_text(size = 10),
        legend.margin=ggplot2::margin(0,0,0,0),
        plot.margin = unit(c(0, 0, 0, 0),"inches"))
g
```

```{r}
png("../figures/gen-mse-comb.png", width = 10, height = 6, units = 'in', res = 300)
g
dev.off()
```

```{r}
tikz(file = "../figures/gen-mse-comb-paper.tex", width = 6, height = 4, symbolicColors = T, colorFileName = "../figures/figcolorsgen.tex")
g
dev.off()
```


## Coverage Table

```{r}
mod.names <- c("$\\tdm$", 
               "$\\tregrct$", 
               "$\\tacwt(\\g)$",
               "$\\tacwt(\\gstar)$",
               "$\\tacwt(\\gstar)$",
               "$\\tacwt(\\gstar)$",
               "$\\tacwt(\\gstar)$",
               "$\\tacwt(\\gstar)$",
               "$\\tacwt(\\syn)$")
xwalk <- data.frame(labs = mod.names, mod = mods)
```


```{r}
coverage.p10long <- coverage.p10 %>%
  mutate(ncov = 10) %>%
  mutate(group = factor(rep(1:20, each=50))) %>%
  select(-sate, -type) %>%
  pivot_longer(!it:group, names_to = "mod", values_to = "val") 


coverage.all <- coverage.p20 %>%
  mutate(ncov = 20) %>%
  mutate(group = factor(rep(1:20, each=50))) %>%
  select(-sate, -type) %>%
  pivot_longer(!it:group, names_to = "mod", values_to = "val") %>%
  rbind(coverage.p10long)


coverage.group <- coverage.all %>%
  group_by(ncov, group, mod) %>%
  summarize(covg = mean(val, na.rm = T)) %>%
  group_by(ncov, mod) %>%
  summarize(sim_se_covg = sd(covg)/sqrt(20),
            group_covg = mean(covg))

coverage.tab <- coverage.all %>%
  group_by(ncov, mod) %>%
  summarize(covg = mean(val, na.rm = T)) %>%
  left_join(coverage.group, by = c("ncov", "mod")) %>%
  mutate(covg = sprintf("%1.2f (%1.3f)", covg, sim_se_covg)) %>%
  select(ncov, mod, covg) %>%
  pivot_wider(names_from = ncov, values_from = covg,
              names_glue = "{.value}{ncov}") %>%
  filter(mod %in% mods) %>%
  mutate(mod = factor(mod, levels = mods)) %>%
  left_join(xwalk, by = "mod") %>%
  mutate(ep = case_when(str_detect(mod, "epsilon") ~ str_extract(mod, "\\d+"),
                                   TRUE ~ "-")) %>%
  arrange(factor(mod, levels = mods))

coverage.tab
```


```{r}
coverage.tab %>%
  select(labs, ep, covg10, covg20) %>%
  kable(booktabs = T, col.names = c("Estimate of the PATE","$\\epsilon$" ,"p = 10", "p = 20"),
                    escape = FALSE, align = "lccc", full_width = F) %>% 
  kable_styling(latex_options = c("hold_position","scale_down")) %>%
  row_spec(row = 0, align = "c", bold=T) %>%
  pack_rows("RCT data only", 1,2) %>%
  pack_rows("Includes auxiliary data", 3,9) %>%
  add_header_above(c(" " = 2, "Coverage" = 2),  bold = T)
```





```{r}
df <- coverage.tab %>%
  select(labs, ep, covg10, covg20) %>%
  kable(booktabs = T, col.names = c("Estimate of the PATE","$\\epsilon$" ,"p = 10", "p = 20"),
                    escape = FALSE, align = "lccc", full_width = F, format = "latex") %>% 
  kable_styling(latex_options = c("hold_position","scale_down")) %>%
  row_spec(row = 0, align = "c", bold=T) %>%
  pack_rows("RCT data only", 1,2) %>%
  pack_rows("Includes auxiliary data", 3,9) %>%
  add_header_above(c(" " = 2, "Coverage" = 2),  bold = T)

write_lines(df, '../figures/gen-coverage.tex')
out <- read_lines('../figures/gen-coverage.tex')
write_lines(out[-c(1, length(out))], '../figures/gen-coverage.tex')
```

## Other Utility Metrics
```{r}
meanslist <- list("10" = auxmeans.p10,"20" = auxmeans.p20)
```

```{r}
auxmean.mse.all <- foreach(p = names(meanslist), .combine = rbind) %do% {
  mean.mse <- meanslist[[p]] %>%
    rownames_to_column(var = "var") %>%
    mutate(it = case_when(str_detect(var, "y") & str_detect(var, "\\d") ~ str_extract(var, "\\d+"),
                          str_detect(var, "y") ~ "0",
                          TRUE ~ NA)) %>%
    fill(it) %>%
    mutate(across(!it & !var, .fns = ~(.x - m.true)^2,.names = "{.col}_err")) %>%
    group_by(it) %>%
    summarize(across(ends_with("err"), ~(sqrt(mean(.x))))) %>%
    pivot_longer(!it, names_to = "privmethod", values_to = "mse") %>%
    group_by(privmethod) %>%
    summarize(e_mse = mean(mse), se_mse = sd(mse)/sqrt(1000)) %>%
    mutate(privmethod = str_replace(privmethod, "\\_err", "")) %>%
    filter(privmethod != "m.true") %>%
    mutate(p = p)
}
```


```{r}
methorder <- c("m.aux", "m.syn", "m.epsilon1","m.epsilon3","m.epsilon6","m.epsilon15","m.epsilon30")
dat.names <- c("$\\g$ (no privacy)",
               "$\\syn$",
               "$\\gstar$",
               "$\\gstar$",
               "$\\gstar$",
               "$\\gstar$",
               "$\\gstar$")
xwalk <- data.frame(labs = dat.names, privmethod = methorder)
```

```{r}
meanmse.tab <- auxmean.mse.all %>%
  filter(privmethod %in% methorder) %>%
  mutate(mse =sprintf("%1.2f (%1.3f)", e_mse*100, se_mse*100)) %>%
  pivot_wider(names_from = p, values_from = c(e_mse, se_mse, mse),
              names_glue = "{.value}_p{p}") %>%
  left_join(xwalk, by = "privmethod")  %>%
  mutate(privmethod = factor(privmethod, levels = methorder)) %>%
  arrange(privmethod) %>%
  mutate(ep = case_when(str_detect(privmethod, "epsilon") ~ str_extract(privmethod, "\\d+"),
                                   TRUE ~ "-"))
```

```{r}
meanmse.tab %>%
  select(labs, ep, mse_p10, mse_p20) %>%
  kable(booktabs = T, col.names = c("Data Release","$\\epsilon$" ,"p = 10", "p = 20"),
                    escape = FALSE, align = "lccc", full_width = F) %>% 
  kable_styling(latex_options = c("hold_position","scale_down")) %>%
  row_spec(row = 0, align = "c", bold=T) %>%
  add_header_above(c(" " = 2, "RMSE of Column Means (x 100)" = 2),  bold = T)
```
```{r}
df <- meanmse.tab %>%
  select(labs, ep, mse_p10, mse_p20) %>%
  kable(booktabs = T, col.names = c("Data Release","$\\epsilon$" ,"p = 10", "p = 20"),
                    escape = FALSE, align = "lccc", full_width = F, format = "latex") %>% 
  kable_styling(latex_options = c("hold_position","scale_down")) %>%
  row_spec(row = 0, align = "c", bold=T) %>%
  add_header_above(c(" " = 2, "RMSE of Column Means (x 100)" = 2),  bold = T)

write_lines(df, '../figures/gen-mean-rmse.tex')
out <- read_lines('../figures/gen-mean-rmse.tex')
write_lines(out[-c(1, length(out))], '../figures/gen-mean-rmse.tex')
```
